(Feat): Generate python stubs in uploaded wheels (#142)
Use py3o_stub_gen to generate stubs/types for python wheels. The stubs right now are not as comprehensive as I would like, but I'll add more as a follow-up/as needed.

Tested that I was successfully able to retrieve these locally on Mac and Linux (and will be testing in CI shortly).