[TRUNK-13581] unified env+repo validation (#202)
Part of [TRUNK-13581](https://linear.app/trunk/issue/TRUNK-13581/move-env-parsing-and-validation-out-of-glue-and-into-dagster)

(apologies in advance for the large diff)

Adds a new 'meta context' validator that validates the combination of an `env::parser::CIInfo` and a `repo::BundleRepo` to prepare for usage in the ETL

The motivation for this is that, there are some instances where we rely on and validate against a combination of values provided by environment variables and the repo details we collect. For example, for many values, if absent from the environment variables, we [fall back to what's in the repo details](https://github.com/trunk-io/trunk/blob/74dc8d323c7a372562ea5d0942d3a63981cf8c65/trunk/py/metrics_data_pipeline/trunk_etl_env_parser/env_parser.py#L114-L118). If we were to slot in the env & repo parser & validator as they are now into the ETL, we could start flagging invalid validation issues for missing env values that would otherwise have fallen back to repo values, and vice-versa. I think this validator also provides a decently-extensible path towards validating the metadata as a whole.

See https://github.com/trunk-io/trunk/pull/19954 for usage

Summary of changes:
 - new `MetaContext` struct & validator
 - changes env parsing to not fail. Previously, inability to detect CI platform or classify a branch name would result in an error. Now allows CI platform and branch class both to be `UNKNOWN`. This mirrors what the ETL is currently doing [(1)](https://github.com/trunk-io/trunk/blob/74dc8d323c7a372562ea5d0942d3a63981cf8c65/trunk/py/metrics_data_pipeline/trunk_etl_env_parser/env_parser.py#L84) [(2)](https://github.com/trunk-io/trunk/blob/74dc8d323c7a372562ea5d0942d3a63981cf8c65/trunk/py/metrics_data_pipeline/trunk_etl_env_parser/env_parser.py#L232) and will continue to allow things like local uploads to be processed. 
 - re-classifies some env & repo validation issues from `INVALID` to `SUBOPTIMAL` since the meta validator will now catch them as `INVALID` (e.g. missing branch name is suboptimal in the isolated contexts of env and repo validation but invalid in the context of unified meta validation)
 - breaks up the `context-py` `test_parse_and_validate.py` into several smaller files since it was getting a bit unwieldy