Swap codeowners parsing order (#222)
Modifies the codeowners parsing logic to first trying parsing a file as a GitLab codeowners file, else try parsing as a GitHub codeowners file.

In integrating with the ETL, we observed GitLab codeowners files being successfully parsed by the _GitHub_ parser, and thus yielding incorrect results with respect to associating owners with a test file. Swapping the order here fixes that. GitLab codeowners syntax is _generally_ a superset of GitHub syntax, so it should be ok if a GitHub codeowners file gets parsed as a _GitLab_ file by our parser so long as it yields the correct results - and it has in local ETL testing.

The relevant diff here is small. The overall diff is inflated due to a test file rename