Remove parsing errors from empty tags and testsuite names (#241)
Baby's first rust PR plz be gentle

## Context
See [here](https://trunk-io.slack.com/archives/C044VBASZ0D/p1734365895854939) for the reasoning behind this. TL:DR is that we were reporting parsing errors for files that don't actually have errors that make them unparsable. These errors were:
1. `StackTraceEmpty` - empty stack trace tag
2. `SystemErrEmpty` empty `<system-err/>`
3. `SystemOutEmpty` empty `system-out/>`
4. `TestSuiteName` empty `testsuite name`

StackAV was getting a ton of errors from this that we were reporting

## Fix
Remove throwing errors relating to these. I left the logic for parsing the tags in there, but just made it so we aren't going to throw an error anymore

## Tests
Update tests in order to addresss