Fix quarantining logic (#315)
When `use-quarantining` is set to true we should defer to the cloud config. When it is set to false we should disable quarantining. Previously we were ignoring the state of that flag and treating everything as though it should be quarantined by always inspecting failures when exit codes were not specified. These exit codes are not specified for the `upload` and `quarantine` commands, so you would essentially always have quarantining enabled when using those commands. In the future this should be renamed and flipped, but this PR updates it to handle those cases appropriately. 

* Add tests to validate upload and quarantine behaviors match expectations