Add reusable test framework for cli tests (#334)
Abstracts out the arg and env orchestration for cli invocation to a builder pattern. The builder fills in default values for parts that are common across existing usages. While we have all of the flags supported by the cli in the builder, we don't have methods for setting all of them because not every flag is actually tested, so those methods would be unused.