Add missing upload parameters (#358)
We were missing the test count and version information.

Test count was caused by us only parsing out the count for junit. Instead we'll allow the uploader specify it. Otherwise we have a circular dependency between upload and test report.

Version information was missing because we weren't updating it when publishing the gem.


<img width="1494" alt="Screenshot 2025-02-05 at 10 34 55 AM" src="https://github.com/user-attachments/assets/f4883e2f-ec27-4dd8-b0c8-fa6336aaba49" />