[TRUNK-14129] Add structured logging
Switches our logging to use the tracing library, which gives us support
for structured logs, using the sentry publisher so that these logs go to
sentry. Adds a single span for execution so we can map out executions as
we encounter issues, and converts our use of error to warn to avoid
spamming oncall.