Fix bazel Bep when using `test` command (#384)
Bazel is generating xml files that we mark as stale when using the `test` subcommand. This works as expected with `upload` because it doesn't check the modification time of the files. Override `exec_start` to be None when using bazel-bep. This is moving that check into bazel since we're treating it as a stronger source of truth.