Fix retries on server errors
Server errors were not being reported as reqwest errors, which meant
that we would try to decode them as json, leading to a decoding error.
Since we don't retry invalid json, this meant we never retried any
server errors. The fix is to use reqest's methods for converting from a
Response with an error code into an error.