Prioritize ci environments (#418)
Prioritizes the ci environments we read in the event there's multiple environments set. If Custom is set, that always wins, if anything else is set, that wins, and only if nothing is set do we go to unknown. If multiple non-Custom environments are set, we choose between them arbitrarily, and make no guarantees about which will be selected (we do guarantee that we'll pick one of them, though).