Add basic config file
Adds a basic config file that fills in args. The file has to be manually
created, and it's applied globally (no multi-repo or multi-org rules),
but it's sufficient as a poc. Also injecting args from the file is done
extremely roughly, as clap doesn't have built-in support for config file
backups.