[codeowners] Improve parsing (#453)
A bit of a code cleanup in code owners around error generation. If one line within a section is invalid, we should not fail to parse the entire code owners files. This update adds a way to identify whether a parsing action is fatal and opts to return an empty vector for those cases. We will still fail if large sections of the code owners are deemed incorrect.