Move away from using quick_junit directly (#490)
Not everything the CLI uses will be directly supported by `quick_junit`. We should move away from assuming the incoming failures are stored in junit and move towards the shared bindings type. This is already how we process this data within our services and this makes the CLI more aligned with it. This is required to support quarantining with rspec and reporting what has been quarantined.