[TRUNK-14636] Allow overriding local repos (#552)
Some customers run tests where they do not check out the repo under test at all in their workflow (an integration test that is testing a backend service by deploying it and then hitting its endpoints, for example). In order to support this, we add the ability to entirely override local repository reading.