Parse internal bin from tarball stream + bindings (#620)
## Why This PR

Similar to `parse_metadata_from_tarball`, this PR aims to introduce a rust method that can be used to extract and parse the internal.bin file from an upload (as opposed to having consumers extract the tarball from the S3 stream themselves). This is because we have a need to get the `internal.bin` file out of uploads in EarlyTL in order to trigger the DevEx AI agent

## Changes

As stated above, this PR adds an additional method + bindings to get and decode the `internal.bin` file out of uploads. There is one thing to note though - unlike the `meta.json` file (which is the first entry in a tarball always) that's **not the case with the internal bin file**. As a result, we end up looking though most of the upload

## Notes

While the bindings are being tested properly, there is no test code in rust so this is only being tested by the bindings test, which isn't ideal. It doesn't look like there's prior work for me to build off of here (that I don't see at least - please correct me if I'm wrong). I'm very out of my element here so if we wanna draw the line now to get tests in for this part, would love some assistance

## Also

Reminder again that I'm very rust basic here - of course please review the PR fully with comments so that I can learn, but plz don't be upset if you see me doing silly things