[TRUNK-14885] Add test_reports argument
Adds a test_reports argument to upload & validate arguments, which
attempts to parse as all valid types of file, falling back to junit if
nothing else works. Since we can't know if an error is because there are
no tests in a bazel file or it's actually invalid, files passed in
test_reports will not raise if they have no test cases.