part 2 update BEP tests to check for resolved status in meta.json (#638)
A lot of tests were not passing the path to the BEP file which made the results incorrect. In addition, I added more checks to API calls and checked that `meta.json` contained a resolved status.

The next PRs will be adding more properties to the `meta.json` that will get used for validation.