part 4 use test runner report for junit report validation (#653)
TRUNK-15041

Split from #634

follow up to #651

Uses the test runner report to override the timestamps that are found in JUnits. Additionally, this will validate the test runner report for various possible glitches (the test runner report comes from Bazel, so it should hopefully be completely valid, but you never know 🤷).