[TRUNK-15076] Support unlimited rows for end output
Superconsole lets you emit rows, in which case they get outputted as a
normal bit of console output, which lets us bypass the issues we were
seeing with long rows and more rows than the tty has visible. Creates an
EndOutput trait we can use for these in order to nicely display the end
reports.