Loading Indicators
Adds the ability to show loading indicators in the superconsole world.
Also extracts out a bunch of the logic for rendering to a separate
thread with channels to send messages so that we can handle messages
from multiple threads (necessary because we have threads spun up to
track progress on api calls).