Add TS Bindings For Getting Both Meta and Internal Bin (#676)
## Context

Philharmonics has needs for both the bundle meta and the internal bin of a bundle at the same time. This PR exposes a binding that grabs both of those while only going over the tar stream at most once. This builds off of code I put in https://github.com/trunk-io/analytics-cli/pull/620

## Tests

Added the respective binding test

## Note

Currently I only added the binding for TS. I'll follow up later with a binding for python - it was giving me a lot of type troubles here (how I was using the `derive` decorators) so electing to solve in its own PR