Stream stdout and stderr (#699)
Bazel outputs its final summary to stdout, but progress logs to stderr. With stdout>stderr, we end up logging the summary above the progress. Even worse, since we pipe it to a variable before printing, we end up sitting there doing nothing while waiting for the subcommand to print, and then dump the whole thing to the screen. This is fine if you're letting it run and not watching, but is a spooky experience if you are watching. Thus, we switch to passing both streams out directly, which also nicely interleaves them the way bazel expects.