[TRUNK-15404] Separate staging notifications from prod
Splits out the staging failures from testing main so that we only send
to production-notifications if we fail to upload to prod, and notify
staging-notifications if we fail to upload to staging. We're doing this
because staging goes down through no fault of its own pretty regularly,
and is creating noise in production-notifications.