Support lazy codeowners (#882)
In order to support lazy codeowners, we need to parse codeowners in a step that does not explode or unnest the bundle_tests column. To do this, adds a variant on the existing parse logic that returns Owners paired with the codeowners file, which we can use to associate the parsed Owners to a test in the lazy frame. We want to live with the eager evaluation for this in particular because the alternative would be re-parsing the same file 100k times, which would be inefficient.