Update context-ruby naming + readme #889
Adds a README for context-ruby that explains how it works and how to build + test it. This also renames it to match the intent. Originally we thought it was going to be bindings in the same way as context-js and context-ruby, but that's not really how it works. Instead it's purely the location of the gem that we release out, so renaming it to match that intent.