Address issues with gen_info_id bindings (#910)
`gen_info_id` has been broken for a while when it comes to variants. The flow was corrected by generate_info_id_variant_wrapper_impl, so update these bindings to use that instead. When updating the bindings in our other repos, we'll need to make sure that the code there isn't already compensating for the double invocation bug.