Don't drop information when validating (#913)
By using quick_junit in the validation, we were throwing away extra information embedded in the BindingsReport. This information included the pre-processed codeowners values. Instead of using quick_junit in validation, accept the BindingsReport directly and store Bindings data directly in the output validated values. This will allow us to continue grbabing `valid_test_suites` and be able to find the extra information we store.